feat: rip out mongo auth
There is no mongo auth. No username, no passwords, no TLS.

It's all gone!

This is getting real now.

In order to do away with the controller node and remove dual writing
we need to get away from state. The only snag is that we can't get
the machine address yet, so we just have a blank one for the controller.

The other issue is that the worker is bouncing because there are
workers looking for the machine.