fix: include IPv6 addresses in controller api addresses
Controller API addresses should include both IPv4 and IPv6 values.

AllMatchingScope only returns SpaceAddresses with an exact ScopeMatch
on the logic in the ScopeMatcher. If the exact ScopeMatch is unavailable,
the next ScopeMatch in preference is returned. However, the current
ScopeMatchPublic matcher differentiates between IPv4 and IPv6 with
preference for IPv4.

Create a new ScopeMatcher to remove the preference for IPv4 addresses
when using the AllMatchingScope method with SpaceAddresses.

This will allow for IPv6 addresses to be written to the controller_api_address
table. They are then returned by to all callers of the various GetAPIAddresses
methods in the controllernode domain.