refactor: rebase storage triggers on latest
Now that the DDL work has been done to add provisioning scope to storage
the storage triggers can be rebased to work on top of this work.

I have added the provisioning scope to storage attachment plan again. I
know that this is only for machine but we should be consistent in
implementation.