fix(removal): remove subordinate units
We had neglected to clear out the unit_principal table when removing
units.

This was fine for regular units, since we dont populate this table, and
principal units, since we block if a unit has subordinates.

However, this caused a FORIEGN KEY contraint error when removing
subordinates.

Include this table in the removal process, and add a test for this case.