chore: add migration token to MigrationSpec
In a previous PR, a token field was introduced for authentication between controllers. The addition of this field to the api client's MigrationSpec object was missed.

This meant that while the controller supports the new field and the RPC params were updated to serialise/deserialise it correctly, anyone using Juju's Go client would be unable to set it.