refactor: differentiate between two controllerNodeChanges channels
The apiAddressSetterWorker's controllerNodeChanges channel has been
renamed to controllerNodeAddressChanges to differentiate it from the
controllerNodeAddressChanges channel in the worker's loop. They are not
the same.

One is for changes to controller nodes themselves, e.g. the list of
controller nodes has changed. The other is for notification if any of
the IP addresses from the known controller nodes change.