feat: implement methods to return API Addresses as HostPorts
GetAPIHostPortsByControllerIDForClients and
GetAPIHostPortsByControllerIDForAgents return slice of HostPorts
where they are grouped by controller ID in ordered
sequence.

Ideally this method wouldn't be necessary, as most of the details
are not available in this domain. However, we are not updating
facade methods at this time and those results include HostPorts
structures.