test: adds DDL constraint test for agent password
It was raised in PR feedback that we are relying on a unique constraint
in the DDL for how the match password implementation works for model
agents.

If this DDL constraint was every change it could possibly introduce a
security problem into the controller. To find this case and raise an
error I have added a unit test that will fail if this case is know
longer true.

This will draw the developers attention to the problem and correctly fix
the issue.