feat: storage provisioning domain service implementation
This commit proposes the new service interface for watching storage
instance entities within the Juju model. The set of watchers and
information getter functions provided exist to replace the mongo based
state calls in the storageprovisioner facade.

To help ease understanding the interface has been broken up into
seperate files for filesystems and volumes. All watchers are concerned
about watching storage instances based on the scope from which they are
provisioned.

Because of the nature of how the facade was implemented for storage
provisioning not all watchers provide the exact data required by the
worker. The facade will be required to do some further leg work to
transform the data.