fix: model machines watcher should watche lifecycle changes
This patch updates the `WatchModelMachines` watcher so that it only reacts
to changes in the lifecycle of the machines, i.e. added machines or
changes in their life.

In order to do this, the custom trigger for entity lifecycles has been
corrected, since it previously only reacted to insertions and deletions,
whilst we should be reacting to insertions and changes in the life.
Also, the watcher has been updated to use this custom trigger (which
already emits the names of the entities) so the mapping from uuid ->
name has been removed. We continue filtering on the name to prevent
seeing container changes.