feat(uniter): implements UpdateNetworkInfo in the uniter
This patch introduces a new method `setUnitRelationNetworks` in
`apiserver/facades/agent/uniter/uniter.go`, replacing the removed `SetUnitRelationNetworks`.
The resolution of networks info is done in the uniter, instead of the domain, because doing it in
the domain would mix up functionnalities from relation domain (fetch relation and set the unit
settings), and from network domain (fetch domain info).

This implementation is not the best, I believe we can go further and factorize some part, because
it call the domain a lot for fetching various uuid and identification information for a lot of
entity (unit, relation, settings).