feat!: remove harvest mode from provisioner task
Following on from the removal from model config, we can also remove
it from the provisioner task. It gets a little hairy because we have
to ensure that we don't break any pre-existing tests. What makes matters
worse is that the default mode for the provisioner task, was harvest
all. This meant that all unknown and destroyed machines where removed.
Lastly, there was a tautalogical test was in one of the tests, that
was returning the wrong instance id, which triggered the unknown test
case, which it should have destroyed mode.