feat: add life to machine instance
machine_removal table is a hack because we haven't correctly modelled
the life of a dying machine correctly. The solution is to add a life
to the machine cloud instance allow it to transition between the states.
Thus we can get rid of the hack.

We have to make this work for 3.6 because it's acting on the
provisioner API, which is an agent facade.

We won't allow the deletion of a machine until the instance has
transitioned to dying.

I've also added a trigger to the common entities to prevent the life
transitioning in reverse. This should prevent the removal worker
breaking based on common held assumptions.