refactor: get type IDs from db for import of devices
The database is the source of truth for type ID to name
mappings. Get the virtual port type and link layer device
type ID to name mappings from the database to use when
inserting new rows during migration import.

This will keep the INSERT query simple as compared to the
alternative of JOINing tables on each insert.