feat: allow migration minion to follow redirects
This change supports migrations to a JAAS environment where the target address is a JIMM controller.
JIMM manipulates the model description before forwarding calls to a backing Juju controller. The migrationmaster should always communicate with JIMM.
When migration minions are moved however, they need to communicate directly with the backing controller.
To do this we return a redirect error from JIMM when an agent tag is presented at login.
The migration minions need to follow this redirect to end up at the backing controller.