feat: adds state layer for storage provisioning
This commit introduces the state layer for storage provisioning. This
initial implementation is concerned with add support for watching
storage entities.

This this support it is possible for the storage provisioner worker to
respond to changes in the model and provision storage as required. This
follows the requirements of what the service layer required.