fix: race condition in provisioner task
This patch fixes a race condition discovered in the provisioner task,
which makes it not-idempotent (at least regarding machine creation).

The symptom of this issue was a machine marked as error status, thus
preventing having the unit placed on it as idle. This behavior was
flaky.
The reason for the failure was that the instance was already started
(error from LXD), and indeed when doing `lxc ls` we could see the
instance correctly started in lxd:
```
Unit             Workload     Agent       Machine  Public address  Ports  Message
dummy-sink/0     waiting      allocating                                  waiting for machine
dummy-source/0*  maintenance  idle                                        Started
juju-qa-test/0*  active       idle                                        hello

Machine  State    Address  Inst id        Base          AZ  Message
0        started           juju-740e47-0  ubuntu@20.04      Container started
1        started           juju-740e47-1  ubuntu@20.04      Container started
2        error             juju-740e47-2  ubuntu@20.04      Failed creating instance record: Instance "juju-740e47-2" already exists
```
And this corresponds to what I saw in the logs:
```
machine-0: 09:40:44 DEBUG juju.worker.provisioner processing machines [0]
machine-0: 09:40:44 DEBUG juju.worker.provisioner processing machines [0]
```
Which indicates that the watcher might be firing twice for the same
machine creation (INSERT) event. This wasn't seen before because we were
using mongodb-based watchers and this behavior might have changed (maybe
a bug in the changestream?).

Now, the real fix is to make the provisioner task robust against this
duplicated events, which makes it idempotent.
The fix is to *only* do start a machine when the machine ID *is not* in
the machineStarting map, which has access protected by a mutex. By
adding this extra step, and also removing the machine ID once we
correctly called StartInstance, the race is fixed.