fix: provisioner task tests races
This commit follows up the previous one, with fixes to tests that
started failing after the provisioner task became idempotent.
These tests were sending two simulated events without synchronizing on
the actual call to StartInstance, and therefore they were taken as
concurrent calls and therefore protected by our mutex. So we never got
the second call to StartInstance.
The fix is to add a channel as sync point and wait until the first
StartInstance gets called to send the second event.