feat: remove resolve constraints
This removes the resolve constraints from state environs policy. We now
have all the information we require to be able to resolve the
constraints correctly. This will need wiring up for all entitites;
machines, applications etc. For now, I've wired it up for the
caas application provisioner.

There is only storage left on the policy and then it's gone. I suspect,
we could just remove it and deal with the consequences.