feat(InstancePoller): cut over to DQLite
I have done this by injecting the domain services directly into the
InstancePoller worker, since it's a controller worker.

This has meant I can remove the instancepoller controller facades and
api code entirely.

There are no changes to the behaviour, with the exception that I have
dropped the call to SetProviderAddress from Mongo and not replaced it.
The reason for this is that setting the provider address is handled by
link layer devices when "SetProviderNetConfig" is called. This could
lead to some breakage in the near future, but will be resolved when we
are more fully cut over to DQLite.