feat: remove machine undertaker
We don't need the machine undertaker, because the removal domain is
going to take the place of the worker. There is no point having a
worker that is advancing a machine removal if we've already trying to
encapsulate it in one worker. This makes it a lot easier to understand
the concept behind it all.