test: add assumption tests for ddl
Adds test to the storage provisioning state layer to check that
assumptions made about the DDL still hold true. This is done to watch
for breakage on the provision scope table with id value changes.

If these tests fail then it means there is a problem that needs to be
fixed in this package.