feat: wire provisioner with machine domain
This patch wires the provisioner facade with the machine domain,
effectively getting rid of the legacy state machine.

For this to be possible, a new method (`IsContainer`) had to be added to
the machine domain (state + service) to replace the legacy
`machine.IsContainer()`. This method simply looks on the machine_parent
table to check if the provided machine is listed as child.

Also, the to be able to get rid of the machines, the call to
`PopulateContainerLinkLayerDevices` from the bridge policy had to be
replaced by the network domain method `DevicesForGuest`.