chore: remove unused unitassigner facades
The unit assigner worker has been removed in a previous patch, now this
patch removes the actual facades and api client.
Although this an agent facade, it can be safely removed without breaking
compatibility with previous version clients. This is because the
unitassigner manifold was only recruited in the jujud-controller
manifolds.