feat: add controller details to ha facade
This is the start of putting back HA. This is a bit of a miss-feature,
but until we have better support for logging in a HA system, this will
have to do. Ideally, we would do this server side and not expose the
controller information like this, but we are where we are.