feat: force state converter to write agent config
This is a bit of a workaround/hack that works around the fact that
we're a controller, that can't connect to dqlite yet - BUT we need
convert the agent into a controller machine. This is the only worker
that must not use the domain services.

There is still an issue about how to correctly insert the controller
nodes, but we can do that later.