docs: remove see more sections
The tutorial frequently linked down to further how-tos and reference.
That was intended as optional "Further reading", but Daniele thinks
it adds bulk and nobody is going to use those links anyway.
I also think it's probably best if the user taking a tutorial never has to leave
the tutorial page.