docs: remove dependency on setup htg and charm-dev blueprint and day 1-n logic
Our tutorial has symmetric setup and teardown sections which strongly recommended
the automatic Multipass charm-dev blueprint path those also covered the manual
+/- Multipass path, all of this content in fact spelled out in a separate how-to.

The charm-dev blueprint is being deprecated, so we need to move away from it.
Daniele thinks mentioning multiple paths creates cognitive overload for the user,
so we should remove or deemphasize that. I also think that, while potentially
useful to a repeat user, that how-to can be quite intimidating, so it's probably
best if we give the instructions directly in the tutorial after all, choosing the
manual path with Multipass. This commit makes all of these changes.

Not being able to hide details with the charm-dev blueprint anymore means that
we have to explain notions such as controller bootstrap quite early. This is however
also an opportunity to make the Deploy section lighter -- it no longer needs to go
over clouds, credentials, and controllers and, if we do the early controller bootstrap
section right, notions such as model, application, unit, and even unit agent and charm
code can be introduced earlier, at least in comments. And the new titles we end up with
in the more gradual setup are an opportunity to move away from the
artificial and external day 0 (plan) - day 1 (deploy) - day n (maintain) logic
and embrace the logic we have in the juju.is/docs happy path: Install Juju,
Connect a cloud, Use charmed applications, with everything we had in day 1 and day n
moved under the latter.