fix: fixes bad cartesian product sql for id
We have sql that attempts to work out either an attached machine or unit
for a volume or filesystem attachment. The code needs to only consider
units when the attachment is not associated with a machine in anyway.

The original sql considered this problem through a no match join but the
problem with this is that the it will match on any machine as a left
join. The better approach is to left join on machine and match on NULL.