feat: add PlaceMachine service method in favor of create machine methods
This patch refactors the machine domain to use the same logic every time
we want to place a new machine. Since the logic is defined in the
machine placement (sub-) domain, we don't allow manual creation of
machines with parents. In any case this wasn't actually needed, because
a machine with a parent is always a container, which is taken care of by
the PlaceMachine method.