fix: allow relation units watcher to work for peer relations
The funtional change is to make a determination at watcher creation,
which endpoint is relevant to an application relation settings change.
If it is a peer relation, the endpoint from the same side is relevant,
otherwise it is the endpoint from the *other* side.

There are a number of attendant refactorings. Because we can get the
information from a single query, we no longer need lazy loading or
transaction wrapping.

To the extent relevant to this change, usage of typed IDs in persistence
concerns has been replaced by strings.