fix(relation): fix `insertRelationUnit` behavior
This patch fixes the behavior of insertRelationUnit. When called with a non-existent relation UUID
or UnitUUID, this function would return a new UUID without any insertion, which can cause bugs.
In such cases, the method should return an error.
It also verifies that if there is not exactly one insert, the method fails (this would mean the SQL
query is wrong or maybe a symptom of data corruption)

Also, it introduces new tests for the `insertRelationUnit` method in the file
`domain/relation/state/relation_test.go`. The following scenarios are addressed:
- Happy path (`TestInsertRelationUnitHappyPath`): Verifies that a unit can be inserted into a
 relation successfully.
- Non-existent relation UUID (`TestInsertRelationUnitRelationUUIDDoesntExist`): Ensures an error is
  returned when the relation UUID does not exist.
- Non-existent unit UUID (`TestInsertRelationUnitUnitUUIDDoesntExist`): Confirms an error is raised
  when the unit UUID does not exist.