feat: wire up api remote callers
This allows the file system to be used in HA. I simplified the
remote objectstore a lot as it was too hard to debug at the moment.
Instead, we just iterate over each remote unitl we find one that
we can use. For file backed HA this can be the MVP and we can optimise
at a later date.

The rest is just the wire up of the api remote callers through the
various manifolds. Also, the api remote callers needs to take the
objectstore services and not the domain services at that's cyclic
dependency matching can't be satisfied.