feat: add machine constraints validation and merge
This patch adds the constraints validation along with the merge of the
provided constraints with the model constraints.

The methods and logic were copied from the application domain, where we
do exactly the same in terms of validation and merge. This was also how
the legacy state worked.

Tests were also copied and updated for the machine domain.