test: add storage provider validator for testing
For some reason the secret domain uses the application service to
construct testing application. Because of this the secret domain needs
to then have all of the dependencies of the application service domain.

This commit adds a test implementation of StorageProviderValidator so
the secret domain does not have to bring in storage packages.