chore: add tests previously commented for container machines
In https://github.com/juju/juju/pull/20024 we had to comment out three
tests in the watcher suite of the machine domain. This was necessary
because the container machine names was going to be addressed in a
follow-up PR. This is the actual change:
https://github.com/juju/juju/pull/20152.

This patch uncomments the tests and makes sure we correctly handle
watchers for containers.