fix: correct order of constraints merge in machine and app domain
The constraints validator of the provider can also merge constraints
with the model constraints.
This patch fixes the order of the passed arguments in the app and
machine domain because the fallback (which should always be the model
constraints) are the first argument, which was inverted.

Tests were fixed accordingly.