refactor: updating endpoint bindings via SetCharm
Replace implementation of updating endpoint bindings in SetCharm with
mergeApplicationEndpointBindings. This allows the logic to be shared
with MergeApplicationEndpointBindings and adds additional validation
for endpoint binding changes on refresh.

There should be no difference between updating endpoint bindings with
the bind or refresh commands. SetCharm can change endpoint bindings even
if the charm is not updated itself.

The extra validation was checking that units of the application have
addresses in the spaces listed for the endpoint bindings updates.