feat: move state serving info to controller
The state serving info from state to the controller. There is a slight
overlap between controller config and state serving info, so we have
to do the right thing and de-dup the information. This requires
changing controller config. Controller config is starting to have it's
own schema, as it should.