feat(status): add ip addresses to machine status for FullStatus
This turned out to be quite fiddly. The proper way of doing this if by
using infra in core/network/address.go to filter, and rank ip addresses.

However, this requires pulling more data out of the database than we
strictly need. Do this for now, but later down the line we might want to
consider refactoring this.