refactor: simplify the objectstore
There is a nasty data race in the objectstore. This is part of the fix,
the other fix is to remove the draining of s3 into the new worker. That
work will be done in a follow up. Once draining is removed, that will
simplify the s3 store dramatically. The file backed one will still be
complex because of the remote retrieval.