fix: incorrect charm storage check
In a previous commit the check to make sure the storage count max for a
charm was a number greater or equal to zero was introduced. This was
done to validate that data coming into the domain was safe.

However we seen that it is valid for a charm to have defined this valud
as -1 to suggest no upper max count on storage exists.