fix(relation/state): endpoint UUID to watch wasn't properly selected in WatchRelatedUnits
The change introduced in #20145 had a small bug.
The endpoint UUID to watch depending on the nature of the relation (peer or not peer) was incorrect.
In a non-peer relation, the excluded endpoint UUID exists on both current relationUnit and any
relationUnit related to another unit of the same application.

This patch fixes the issue by refactoring the selection logic. We store the various endpointUUID
of every relationUnit linked to a relation. If we have only one, it means that we are in a peer
relation, so we watch it. If we have several, we should have 2, so we exclude the one linked to the
current unit and return the other.

- Removed the unused `isPeer` variable.
- Introduced `localEndpointUUID` and refactored endpoint determination to handle local and remote
 endpoints explicitly.
- Ensured correct processing of peer relations by accurately selecting endpoints using the
 `endpointUUIDs` set.

I also re-enable flaky test that was impacted by the bug.