feat: start replacement of GetEnviron in apiserver
Replace the environ used to get lxd profiles with a call to the
machine service to do the work.

Remove use of the environ for storage concerns in provisioninginfo.
Storage will be reworked shortly and is not currently expected to
work.

No unit tests were updated, they currently do not exist.