feat: wire up remove entity watcher
The removal entity watcher, watches for changes to a given entity
and if the life is either dying or dead it will be notified.

This required a few custom triggers so that we can get the uuid vs
the name for all types.

Of interest, this is the way we would create the all watcher, it's
so easy now!