refactor: converges on a single set of data types for application
To create units in the application domain we had two data types that
represented the same information. That was what was required in the
domain to create new units.

This became a problem when talking about storage and what storage needed
to be created for a new unit. To simplify down this task this commit
gets application domain back to one source of truth for how to create a
unit so the task of specifying storage becomes clearer.

This had a wide ranging impact on unit tests in the domain. A lot of
unit tests were concerned with having a unit created but didn't care how
it was made. This presented the oppurtunity to provide a testing helper
that offered simplified this job removing a lot of boiler plate code.