fix: auth func in deployer facades
Currently there is an issue when trying to deploy with a placement
directive on an existant machine. The unit never actually starts and the
deployer logs that getting the life of the unit gets "permission
denied".

The auth function in the deployer facades is wrong. The problem is that
it has stale data because it requests the units for the machine only
once when constructing the API instance.
Instead, this should be a closure.

This patch addresses this issue and therefore the units for the machine
get correctly deployed.