feat: get an environ for finding images and agents
GetBootstrapEnviron returns an environs.Bootstrap for use with calls to
functions in the environs/tools and environs/imagemetadata packaes.
Normally this functionality would be pushed into a service. However
functions in those packages are also used before an instance is created
during bootstrap, thus unable to use a domain method.