refactor(network): remove Rollback logic for network migration
This patch removes the rollback logic for network domain migration since it is not required.
Network domain is a model-related domain, which means that in case of failure, the whole domain DB
is deleted anyway.

See comment of the Rollback method in Operation interface:
core/modelmigration/coordinator.go#L57-L63