refactor: volume id info for filesystems
We have introduced to the storageprovisioning service the ability to get
information about a filesystem in the model. With this information we
also include the volume id of the volume that backs the filesystem.

If the filesystem isn't backed by a volume this id value would be an
empty string. Empty string implying something doesn't exist is a loose
concept and not one that is a concrete point of our data manual.

The better way to represent this information would be with a volume
structure describing the volume that backs the filesystem. We can then
make this structure nil when it doesn't exist. This contract is much
stronger and one that can be validated inline with our DDL.

While here I also fixed an issue with the sql for getting filesystems
that had incorrect joins to the storage instance table. This would
result in a filesystem not found error if a filesystem wasn't associated
with a storage instance.

Have also implemented a regression test to show the issue is fixed and
catch this problem in the future.