fix(network/state): avoid inconsistent subnets resolution while merging addresses
This patch fixes possible data corruption while updating subnet from instance poller.
The actual façade doesn't provide a provider subnet id for each address, but instead, provides
a global provider subnet id for all addresses of the same device.

On LXD, it causes instance poller to overwrite the Ipv6 addresses subnetUUID to link it to the
subnet of the ipv4 address on the same device, which is wrong.

To mitigate this bug without updating the facade, we introduce a subnet verification logic before
updating to ensure IP addresses match their subnet according to CIDR.

This doesn't change any previous unit tested behavior, but a test has been added to validate the
introduced behavior.