chore: remove legacy state Model from modelmanager facades
This patch removes the legacy state.Model() call for the modelmanager
facades.
The main impact is that since one of the facades is DestroyModels(), and
we're not calling (legacy)model.Destroy() anymore, we replaced it with
the removal service.
Since the removal service is not yet fully finished for model destroy
(there are some issues removing the associated machines, for example),
model destroy is not fully functional.