chore(apiserver): remove state controller from apiserversuite
This patch is the first part of chores to remove state from apiserversuite.

It removes the dependency to most state related function, and fix tests that depends on them.

The only remaining dependency is to the StatePool, since it is deeply coupled in several tests and
maybe need more progress on related tasks.

This patch tries to reduce the impact on other chore task by only replacing some state injection by
nil pointer, since the legacy state is a noop.