docs: move tutorial observe section into htg collect metrics about a controller
That section was in the tutorial for 2 reasons:

(1) We were using it to highlight a new feature that was also not exposed
anywhere else in Juju docs.
(2) We were using it to add more substance to the day 2 / maintain section.

However, we now have a how-to guide that can host those steps (and which
was always a bit too abstract without them anyway), so reason (1) is gone.
Also, we have now decided to move away from the day n logic,
so reason (2) is gone as well. Besides, this content was never really a good
 fit for day 2 anyway -- if you wanted to observe, you'd set
up your observability stack on day 1. It was really more of an excuse to
showcase the capabilities of the Juju ecosystem. But, that's something we
show pretty well with the quick Mattermost-PostgreSQL deploy, configure, scale
bit anyway. If we want to highlight even more, we can/should do that in the abstract
-- after all, it's not just COS that's great but also Charmed K8s, IAM, etc.,
and all of those now have pretty decent docs of their own.
Fundamentally, as we concluded in a previous commit, the main thing the Juju tutorial
needs to do is give people a taste of the Juju paradigm, which is fundamentally
Juju - applications (so long as they are charmed) - clouds (so long as you've
connected them to Juju).