feat: file objectstore removal all
This adds the implementation for file objectstore removal code. This
should work in both HA and non-HA modes, as the witness for the object
store should happen everywhere.

The s3 objectstore is a lot more difficult and requires more thought.
It will be wise to wait for[1], which cleans up the s3 objectstore,
which might help the implementation of this.

 1. https://github.com/juju/juju/pull/20222