refactor(apiserver/firewaller): remove legacy `State` dependency
This patch removes the dependency to the legacy `State` interface within the `firewaller` facade.

The logic which was still relying on this was cross-model relation related, and was already stubbed
out through shim and not implemented functions.

This patch pushes it a step further by removing the shim and introduces unexported stub function
`getMacaroon` and `oneSetRelationStatus` in `apiserver/facades/controller/firewaller/firewaller.go`.
Those functions are meant to be implemented or removed whenever crossmodel relation will have a
proper domain.