feat: initial offer ddl
The DDL is very simple, link an offer to different application endpoints
via join table and ensure that all offers have endpoint on the same
application.

There is a referential integrity problem. The offer related tables are
in the model db and permission table for offers is in the controller db.