feat: remove agent tools checker from jujud
This requires it to be run on the controller. Jujud is never going to
run on a controller. We've left that behind now. All machines know
upfront if they're a controller or not. Redundant controller pets are
gone. Everything is cattle.

This will then allow us to not have to use the API facade to check
the agent tools version. Thus, we can remove another controller facade
in the future.