feat: remove the model deletion job
We can't have the model deletion job in the removal worker. The
removal worker lives inside of the model worker manager and that
starts to tear itself down once the model becomes dead. This prevents
the removal job from continuing on. It's dead.

The last thing the removal worker can do is set the model to dead.
Everything after that is a hope and a prayer.