feat: watch units on a machine
There is a logical race between the removal of a machine and a unit.
If the machine removal job is called before all unit removal jobs
and the machiner recieves no new messages, then we won't pick up the
fact that the machine should die.

This is a massive pain, as we're not at liberty to change the machiner
without being backwards compatible with 3.6.

Instead, the watcher in the machiner will also watch for units dying
and will notify based on that. This requires some funky custom trigger
to get it to work. This works when on inserting or deleting a unit
or updating the unit life, then distpatching the unit notification on
a custom topic. The watcher looks for the machine name as a predicate
and rejects anything not matching that.

Triggers with changestream subscriptions are VERY powerful. This
would be a nightmare done purely in code.

We don't care about k8s in this scenario as the machiner is never
run on k8s ;-)