feat: add public address found by provider
Public addresses are not see by the machine itself, so add them to
existing devices if not seen yet. Subnet suffixes are added to any
address value without it. The suffix is first found via the subnet CIDR
if provided in input, then from the subnet itself, lastly 32, or 128 is
applied if IPv4 or IPv6.

Beware of captialization when comparing incoming and exising data in
MergeLinkLayerDevice. Use lower case in normalized data to ensure
matching.

Drive by: Do not try to update the origin of addresses which are already
machine typed and not seen by the provider. Otherwise we'll attempt work
on all loopback devices every time MergeLinkLayerDevice runs.