fix: prevent check reporting wrong life
Calling Check will always return the original life value **EVEN** if
the life of the model has changed. This is wrong in 3.6 and in the
life flag watcher. This can lead to strange errors for model removal.
If the life has changed, and just before we bounce, something checks
if the life has changed, it will return the invalid value.

This is a nightmare to track down.