fix: return specfic error on dispatch failure
If the subscription is being unsubscribed during a dispatch return the
error. This allows us to trap that in the event multiplexer.

The event multiplexer now handles the sentinel error and logs out the
information. We never want to stop the dispatching of other
subscriptions. This is why sometimes you'd loose events during the
day to day of operations. We now guarantee that every event in a
dispatch set will at least be sent to the subscription. If the
subscription is being unsubscribed during the dispatch then it wouldn't
be passed on, but the guarantee still holds. The subscription was
given the event.